All Done!

+++ src/main/java/com/game/card/service/Impl/CardServiceImpl.java
@@ -2,7 +2,7 @@ package com.game.card.service.Impl;
 
 
 import com.game.card.dto.DtoCard;
-import com.game.card.dto.DtoUpdateCard;
+import com.game.card.dto.dtoRequest.DtoUpdateCard;
 import com.game.card.mapper.CardMapper;
 import com.game.card.model.Card;
 import com.game.card.model.Language;
@@ -25,10 +25,13 @@ public class CardServiceImpl implements CardService {
 
     private final CardMapper cardMapper;
 
+
     public CardServiceImpl(CardRepository cardRepository, LanguageRepository languageRepository, CardMapper cardMapper) {
+
         this.cardRepository = cardRepository;
         this.languageRepository = languageRepository;
         this.cardMapper = cardMapper;
+
     }
 
 
@@ -37,11 +40,16 @@ public class CardServiceImpl implements CardService {
 
         Card card = cardMapper.toCard(dtoCard);
         Optional<Language> language = languageRepository.findByLanguageName(dtoCard.getLanguage());
+
         if (language.isEmpty()) {
+
             throw new EntityNotFoundException("Language not found");
+
         }
+
         card.setLanguage(language.get());
         cardRepository.save(card);
+
         return dtoCard;
 
     }
@@ -52,6 +60,7 @@ public class CardServiceImpl implements CardService {
         Card card = cardRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Card not found"));
         DtoCard dtoCard = cardMapper.toDtoCard(card);
         dtoCard.setLanguage(card.getLanguage().getName());
+
         return dtoCard;
 
     }
@@ -59,7 +68,17 @@ public class CardServiceImpl implements CardService {
     @Override
     public void deleteCardById(Long id) {
 
-        cardRepository.deleteById(id);
+        Optional<Card> cardOptional = cardRepository.findById(id);
+
+        if (cardOptional.isEmpty()) {
+
+            throw new EntityNotFoundException("Card not found");
+
+        } else {
+
+            cardRepository.deleteById(id);
+
+        }
 
     }
 
@@ -70,10 +89,13 @@ public class CardServiceImpl implements CardService {
         List<DtoCard> response = new ArrayList<>();
 
         for (Card card : cardList) {
+
             DtoCard dtoCard = cardMapper.toDtoCard(card);
             dtoCard.setLanguage(card.getLanguage().getName());
             response.add(dtoCard);
+
         }
+
         return response;
 
     }
@@ -83,11 +105,15 @@ public class CardServiceImpl implements CardService {
 
         List<Card> cardList = cardRepository.findCardByLanguage(language);
         List<DtoCard> response = new ArrayList<>();
+
         cardList.forEach(card -> {
+
             DtoCard dtoCard = cardMapper.toDtoCard(card);
             dtoCard.setLanguage(card.getLanguage().getName());
             response.add(dtoCard);
+
         });
+
         return response;
 
     }
@@ -96,15 +122,23 @@ public class CardServiceImpl implements CardService {
     public DtoCard updateCard(Long id, DtoUpdateCard dtoCardUpdate) {
 
         Optional<Card> cardOptional = cardRepository.findById(id);
+
         if (cardOptional.isEmpty()) {
+
             throw new EntityNotFoundException("Card not found");
+
         } else {
+
             Card cardToUpdate = cardOptional.get();
+
             cardToUpdate.setTranslation(dtoCardUpdate.getTranslation());
             cardToUpdate.setWord(dtoCardUpdate.getWord());
+
             cardRepository.save(cardToUpdate);
+
             DtoCard dtoCard = cardMapper.toDtoCard(cardToUpdate);
             dtoCard.setLanguage(cardToUpdate.getLanguage().getName());
+
             return dtoCard;
         }
